Show helpful admin command error if it's a IrcProvisioningError (#1702)
* Check if instance of IrcProvisioningError

* changelog

* Fix a typing bug.

+++ src/bridge/AdminRoomHandler.ts
@@ -27,6 +27,7 @@ import * as RoomCreation from "./RoomCreation";
 import { getBridgeVersion } from "matrix-appservice-bridge";
 import { IdentGenerator } from "../irc/IdentGenerator";
 import { Provisioner } from "../provisioning/Provisioner";
+import { IrcProvisioningError } from "../provisioning/Schema";
 
 const log = logging("AdminRoomHandler");
 
@@ -239,6 +240,9 @@ export class AdminRoomHandler {
         }
         catch (ex) {
             log.error(`Failed to handle !plumb command:`, ex);
+            if (ex instanceof IrcProvisioningError) {
+                return new MatrixAction(ActionType.Notice, `Failed to plumb room. ${ex.message}`);
+            }
             return new MatrixAction(ActionType.Notice, "Failed to plumb room. Check the logs for details.");
         }
         return new MatrixAction(ActionType.Notice, "Room plumbed.");
@@ -272,6 +276,9 @@ export class AdminRoomHandler {
         }
         catch (ex) {
             log.error(`Failed to handle !unlink command:`, ex);
+            if (ex instanceof IrcProvisioningError) {
+                return new MatrixAction(ActionType.Notice, `Failed to plumb room. ${ex.message}`);
+            }
             return new MatrixAction(ActionType.Notice, "Failed to unlink room. Check the logs for details.");
         }
         return new MatrixAction(ActionType.Notice, "Room unlinked.");
@@ -282,7 +289,7 @@ export class AdminRoomHandler {
         // check that the server exists and that the user_id is on the whitelist
         const ircChannel = args[0];
         const key = args[1]; // keys can't have spaces in them, so we can just do this.
-        let errText = null;
+        let errText: string|null = null;
         if (!ircChannel || !ircChannel.startsWith("#")) {
             errText = "Format: '!join irc.example.com #channel [key]'";
         }
